fix panel bug
I created a temporary fix for the accordion/panel components. The problem had to do with the displayMode in the panelComponent being initialized to one value and the displayMode in the accordion being different than that value. The problem was manifesting itself through the HostBindings in the panel. The temporary fix was to eliminate the displayMode input from the accordion completely and just set the input on the Panel manually.